pyproject.toml
[project]
name = "python-etl-sql-postgres"
version = "0.1.0"
description = "Python Demo ETL Pipeline: SQL Server to PostgreSQL"
authors = [
    {name = "kimberly-emerson", email = "kimberly.emerson@outlook.com"},
]
dependencies = [
    "psycopg2>=2.9.10",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "pandas>=2.3.3",
    "black>=25.9.0",
    "argparse>=1.4.0",
    "python-decouple>=3.8",
]
requires-python = "==3.14.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
